dom/webgpu/mochitest/test_too_many_color_attachments.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>Test for Bug 1772557</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css" />
    <script>
      ok(
        SpecialPowers.getBoolPref("dom.webgpu.enabled"),
        "Pref should be enabled."
      );

      add_task(too_many_color_attachments);

      async function too_many_color_attachments() {
        const adapter = await navigator.gpu.requestAdapter({});
        const device = await adapter.requestDevice({});
        device.pushErrorScope("validation");
        const texture = device.createTexture({
          size: [60, 37, 238],
          format: "rg16sint",
          usage: GPUTextureUsage.RENDER_ATTACHMENT,
        });
        const view = texture.createView({});
        const encoder = device.createRenderBundleEncoder({
          colorFormats: [
            "rg8uint",
            "rg8uint",
            "rg8uint",
            "rg8uint",
            "rg8uint",
            "rg8uint",
            "rg8uint",
            "rg8uint",
            "rg8uint",
          ],
        });
        const renderBundle = encoder.finish({});
        const commandEncoder = device.createCommandEncoder({});
        const renderPassEncoder = commandEncoder.beginRenderPass({
          colorAttachments: [
            {
              view,
              loadOp: "load",
              storeOp: "store",
            },
          ],
        });
        renderPassEncoder.executeBundles([renderBundle]);
        renderPassEncoder.end();
        await device.popErrorScope();
        ok(true, "test completed without crashing");
      }
    </script>
  </head>
  <body>
    <a
      target="_blank"
      href="https://bugzilla.mozilla.org/show_bug.cgi?id=1772557"
      >Mozilla Bug 1772557</a
    >
    <p id="display"></p>
    <div id="content" style="display: none"></div>
    <pre id="test"></pre>
  </body>
</html>
